add tabs options
