+++ Conexa/Views/UsersViews.swift
@@ -1,0 +1,15 @@
+//
+//  UsersViews.swift
+//  Conexa
+//
+//  Created by Matias Palmieri on 27/05/2024.
+//
+
+import SwiftUI
+
+struct UsersViews: View {
+    
+    var body: some View {
+        Text("Users view")
+    }
+}
